requirements.txt
cycler==0.10.0
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
progressbar==2.5
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
